requirements.txt
https://github.com/HuyaneMatsu/hata/archive/master.zip
python-dotenv